Copy over README and CHANGELOG
