+++ easy-ui-react/package.json
@@ -17,7 +17,7 @@
     "directory": "dist"
   },
   "scripts": {
-    "build": "run-p build:js build:types",
+    "build": "rm -rf dist && run-p build:js build:types",
     "build:js": "vite build",
     "build:types": "tsc --build",
     "clean": "rm -rf node_modules dist .turbo",
